Crate vk_generator

Automatically generates bindings for the Vulkan API. In order to use, first create a VkRegistry struct with a valid Vulkan XML file, and then call either one of the two generation functions (gen_global() or gen_struct()) or a custom one. Further documentation can be found in VkRegistry and GenConfig, and looking over the provided examples is encouraged.

Structs

GenConfig
Configuration options fot the Vulkan generator

VkRegistry
A struct representation of the Vulkan XML registry.

VkVersion